+++ bin/maze-runner
@@ -63,6 +63,9 @@ require_relative '../lib/maze/plugins/bugsnag_reporting_plugin'
 require_relative '../lib/maze/plugins/cucumber_report_plugin'
 require_relative '../lib/maze/plugins/global_retry_plugin'
 
+# Require monkey-patches after everything else
+require_relative '../lib/utils/selenium_money_patch'
+
 # Encapsulates the MazeRunner entry point
 class MazeRunnerEntry
