+++ lib/maze/capabilities.rb
@@ -22,53 +22,7 @@ module Maze
         capabilities
       end
 
-
-
-      # selenium jsonwp
-      #
-      # capabilities = Selenium::WebDriver::Remote::Capabilities.new
-      # capabilities['platform'] = 'Linux'
-      # capabilities['osVersion'] = '18.04'
-      # capabilities['browserName'] = 'firefox'
-      # capabilities['version'] = '104'
-      # capabilities['resolution'] = '2560x1920'
-      # capabilities['bitbar_apiKey'] = '<insert your BitBar API key here>'
-
-# selenium w3c
-
-      # capabilities = Selenium::WebDriver::Remote::Capabilities.new
-      # 'platformName' => 'Linux',
-      # capabilities['browserName'] = 'firefox'
-      # capabilities['browserVersion'] = '104'
-      # capabilities['bitbar:options'] = {
-      #   'apiKey' = '<insert your BitBar API key here>'
-      # 'resolution' = '2560x1920'
-      # 'osVersion' = '18.04'
-      # }
-
-      # appium jsonwp
-      #
-      # desired_capabilities_cloud = {
-      #   'bitbar_apiKey' => '<insert your BitBar API key here>',
-      #   'bitbar_device' => 'Asus Google Nexus 7 2013 6.0.1',
-      #   'platformName' => 'Android',
-      #   'deviceName' => 'Android Phone',
-      #   'automationName' => 'Appium',
-      # }
-
-      # browserstack (w3c)
-      #
-      # 'platformName' => 'android',
-      #   'platformVersion' => version,
-      #   'deviceName' => device,
-      #   'autoGrantPermissions' => 'true',
-      #   'bstack:options' => {
-      #     "appiumVersion" => appium_version,
-      #   },
-
-
-
-# @param browser_type [String] A key from @see browsers_bb.yml
+      # @param browser_type [String] A key from @see browsers_bb.yml
       # @param local_id [String] unique key for the SB tunnel instance
       # @param capabilities_option [String] extra capabilities provided on the command line
       def for_bitbar_browsers(browser_type, api_key, local_id, capabilities_option)
