+++ lib/maze/hooks/appium_hooks.rb
@@ -1,25 +1,4 @@
 # Contains logic for the Cucumber hooks when in Appium mode
-
-module Selenium
-  module WebDriver
-    module Error
-      class ServerError < StandardError
-        def initialize(response)
-          if response.is_a? String
-            super(response)
-          else
-            $logger.error "Server response: #{response.inspect}"
-            super("status code #{response.code}")
-          end
-        end
-      end # ServerError
-    end # Error
-  end # WebDriver
-end # Selenium
-
-
-
-
 module Maze
   module Hooks
     # Hooks for Appium mode use
@@ -143,7 +122,6 @@ module Maze
                                                               config.os_version,
                                                               config.capabilities_option
           capabilities['bitbar:options']['app'] = config.app
-          # capabilities['appium:bundleId'] = config.app_bundle_id
         end
 
         $logger.info "Capabilities: #{capabilities.inspect}"
